+++ TestApplication/Source/TestScene103.h
@@ -1,6 +1,6 @@
 // =============================================================================
 // @AUTHOR Vik Pandher
-// @DATE 2025-11-17
+// @DATE 2026-01-12
 
 #pragma once
 
@@ -207,12 +207,12 @@ protected:
     unsigned int uiCamera_EntityId_;
 
     unsigned int cursor_EntityId_;
-    unsigned int cursorPositionRenderedMeshIndex_;
-    unsigned int cursorPressRenderedMeshIndex_;
-    unsigned int cursorReleaseRenderedMeshIndex_;
-    unsigned int cursorPositionCollisionPointIndex_;
-    unsigned int cursorPressCollisionPointIndex_;
-    unsigned int cursorReleaseCollisionPointIndex_;
+    size_t cursorPositionRenderedMeshIndex_;
+    size_t cursorPressRenderedMeshIndex_;
+    size_t cursorReleaseRenderedMeshIndex_;
+    size_t cursorPositionCollisionPointIndex_;
+    size_t cursorPressCollisionPointIndex_;
+    size_t cursorReleaseCollisionPointIndex_;
     static const unsigned int s_cursorPositionCollisionShapeId_ = 100;
     static const unsigned int s_cursorPressCollisionShapeId_ = 101;
     static const unsigned int s_cursorReleaseCollisionShapeId_ = 102;
